move useUsers.ts into useQuery directory
