+++ frontend/src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { default as useMultipleSelect } from './useMultipleSelect'
-export { default as useUsers } from './useUsers'
-export * from './useUsers'
+export { default as useUsers } from './useQuery/useUsers'
+export * from './useQuery/useUsers'
 export { default as useDidUpdateEffect } from './useDidUpdateEffect'
